+++ lib/app/pages/features/diet/pages/nutrients_tab_page.dart
@@ -95,7 +95,6 @@ class NutrientsTabPage extends StatelessWidget {
                           ),
                         ],
                       ),
-                      const SizedBox(height: 50),
                     ],
                   ),
                 ),
