+++ lib/app/pages/features/motivation_quotes/pages/widgets/random_quote_widget.dart
@@ -1,0 +1,65 @@
+import 'package:fitness_prodigy/app/core/constants.dart';
+import 'package:fitness_prodigy/app/pages/features/motivation_quotes/cubit/motivation_quotes_cubit.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:google_fonts/google_fonts.dart';
+
+class RandomQuoteButton extends StatelessWidget {
+  const RandomQuoteButton({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return InkWell(
+      splashColor: Colors.transparent,
+      onTap: () {
+        context.read<MotivationQuotesCubit>().getRandomQuote();
+      },
+      child: Padding(
+        padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 10) +
+            const EdgeInsets.only(top: 20),
+        child: Container(
+            padding: const EdgeInsets.all(15),
+            decoration: const BoxDecoration(
+              gradient: LinearGradient(
+                colors: kHomeGradient,
+                begin: Alignment.topCenter,
+                end: Alignment.bottomCenter,
+              ),
+              borderRadius: BorderRadius.all(
+                Radius.circular(20),
+              ),
+              boxShadow: [
+                BoxShadow(
+                  color: Colors.black26,
+                  blurRadius: 5,
+                  offset: Offset(0, 10),
+                ),
+              ],
+            ),
+            child: Row(
+              children: [
+                const Icon(
+                  Icons.all_inclusive,
+                  color: Colors.amber,
+                  size: 30,
+                ),
+                Expanded(
+                  child: Align(
+                    alignment: Alignment.center,
+                    child: Text(
+                      'RANDOM QUOTE',
+                      style: GoogleFonts.montserrat(
+                          fontSize: 18,
+                          color: Colors.white,
+                          fontWeight: FontWeight.w700),
+                    ),
+                  ),
+                ),
+              ],
+            )),
+      ),
+    );
+  }
+}
